+++ crates/sncast/src/lib.rs
@@ -55,6 +55,7 @@ pub type NestedMap<T> = HashMap<String, HashMap<String, T>>;
 pub enum AccountType {
     #[serde(rename = "open_zeppelin")]
     OpenZeppelin,
+    Argent,
     Ready,
     Braavos,
 }
@@ -65,8 +66,8 @@ impl FromStr for AccountType {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "open_zeppelin" | "open-zeppelin" | "oz" => Ok(AccountType::OpenZeppelin),
-            // We still accept "argent" for backward compatibility
-            "argent" | "ready" => Ok(AccountType::Ready),
+            "argent" => Ok(AccountType::Argent),
+            "ready" => Ok(AccountType::Ready),
             "braavos" => Ok(AccountType::Braavos),
             account_type => Err(anyhow!("Invalid account type = {account_type}")),
         }
@@ -75,7 +76,11 @@ impl FromStr for AccountType {
 
 impl Display for AccountType {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{self:?}")
+        if self == &AccountType::Ready {
+            write!(f, "{:?}", AccountType::Argent)
+        } else {
+            write!(f, "{self:?}")
+        }
     }
 }
 
@@ -386,7 +391,7 @@ pub fn get_account_data_from_keystore(
         .and_then(|account_type| account_type.parse().ok());
 
     let public_key = match account_type.context("Failed to get type key")? {
-        AccountType::Ready => parse_to_felt("/variant/owner"),
+        AccountType::Argent | AccountType::Ready => parse_to_felt("/variant/owner"),
         AccountType::OpenZeppelin => parse_to_felt("/variant/public_key"),
         AccountType::Braavos => get_braavos_account_public_key(&account_info)?,
     }