+++ crates/sncast/src/starknet_commands/account/create.rs
@@ -7,6 +7,8 @@ use camino::Utf8PathBuf;
 use clap::Args;
 use console::style;
 use conversions::IntoConv;
+use foundry_ui::UI;
+use foundry_ui::components::warning::WarningMessage;
 use serde_json::json;
 use sncast::helpers::braavos::{BraavosAccountFactory, check_braavos_account_compatibility};
 use sncast::helpers::constants::{
@@ -63,7 +65,16 @@ pub async fn create(
     provider: &JsonRpcClient<HttpTransport>,
     chain_id: Felt,
     create: &Create,
+    ui: &UI,
 ) -> Result<AccountCreateResponse> {
+    // TODO: Remove this warning once we drop Argent account type
+    if create.account_type == AccountType::Argent {
+        ui.println(&WarningMessage::new(
+            "Argent has rebranded as Ready. The `--argent` option for `account create` is now deprecated, please use `--ready` instead.",
+        ));
+        ui.print_blank_line();
+    }
+
     // Braavos accounts before v1.2.0 are not compatible with starknet >= 0.13.4
     // For more, read https://community.starknet.io/t/starknet-devtools-for-0-13-5/115495#p-2359168-braavos-compatibility-issues-3
     if let Some(class_hash) = create.class_hash {
@@ -73,7 +84,7 @@ pub async fn create(
     let salt = extract_or_generate_salt(create.salt);
     let class_hash = create.class_hash.unwrap_or(match create.account_type {
         AccountType::OpenZeppelin => OZ_CLASS_HASH,
-        AccountType::Ready => READY_CLASS_HASH,
+        AccountType::Argent | AccountType::Ready => READY_CLASS_HASH,
         AccountType::Braavos => BRAAVOS_CLASS_HASH,
     });
     check_class_hash_exists(provider, class_hash).await?;
@@ -171,7 +182,7 @@ async fn generate_account(
                 OpenZeppelinAccountFactory::new(class_hash, chain_id, signer, provider).await?;
             get_address_and_deployment_fee(factory, salt).await?
         }
-        AccountType::Ready => {
+        AccountType::Argent | AccountType::Ready => {
             let factory =
                 ArgentAccountFactory::new(class_hash, chain_id, None, signer, provider).await?;
 
@@ -268,11 +279,11 @@ fn create_to_keystore(
                 }
             })
         }
-        AccountType::Ready => {
+        AccountType::Argent | AccountType::Ready => {
             json!({
                 "version": 1,
                 "variant": {
-                    "type": AccountType::Ready,
+                    "type": format!("{account_type}"),
                     "version": 1,
                     "owner": format!("{:#x}", private_key.verifying_key().scalar()),
                     "guardian": "0x0",