+++ crates/sncast/src/starknet_commands/account/deploy.rs
@@ -208,7 +208,7 @@ async fn get_deployment_result(
     ui: &UI,
 ) -> Result<InvokeResponse> {
     match account_type {
-        AccountType::Ready => {
+        AccountType::Argent | AccountType::Ready => {
             let factory = ArgentAccountFactory::new(
                 class_hash,
                 chain_id,
@@ -392,7 +392,7 @@ pub(crate) fn compute_account_address(
     chain_id: Felt,
 ) -> Felt {
     match account_type {
-        AccountType::Ready => get_contract_address(
+        AccountType::Argent | AccountType::Ready => get_contract_address(
             salt,
             class_hash,
             &[private_key.verifying_key().scalar(), Felt::ZERO],