+++ crates/sncast/tests/e2e/account/create.rs
@@ -8,7 +8,9 @@ use crate::helpers::env::set_create_keystore_password_env;
 use camino::Utf8PathBuf;
 use conversions::string::IntoHexStr;
 use serde_json::{json, to_string_pretty};
-use shared::test_utils::output_assert::{assert_stderr_contains, assert_stdout_contains};
+use shared::test_utils::output_assert::{
+    assert_stderr_contains, assert_stdout, assert_stdout_contains,
+};
 use snapbox::assert_matches;
 use sncast::AccountType;
 use sncast::helpers::constants::{
@@ -76,6 +78,72 @@ pub async fn test_happy_case(account_type: &str) {
                     "public_key": "0x[..]",
                     "salt": "0x1",
                     "type": get_formatted_account_type(account_type)
+                }
+            }
+        }
+    );
+
+    assert_matches(to_string_pretty(&expected).unwrap(), contents);
+}
+
+// TODO: Remove this test once we drop Argent account type
+#[tokio::test]
+pub async fn test_happy_case_argent_with_deprecation_warning() {
+    let temp_dir = tempdir().expect("Unable to create a temporary directory");
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--accounts-file",
+        accounts_file,
+        "account",
+        "create",
+        "--url",
+        URL,
+        "--name",
+        "my_account",
+        "--salt",
+        "0x1",
+        "--type",
+        "argent",
+    ];
+
+    let snapbox = runner(&args).current_dir(temp_dir.path());
+    let output = snapbox.assert().success();
+
+    assert_stdout(
+        output,
+        indoc! {r"
+        [WARNING] Argent has rebranded as Ready. The `--argent` option for `account create` is now deprecated, please use `--ready` instead.
+        
+        Success: Account created
+
+        Address: 0x0[..]
+        
+        Account successfully created but it needs to be deployed. The estimated deployment fee is [..] STRK. Prefund the account to cover deployment transaction fee
+
+        After prefunding the account, run:
+        sncast --accounts-file accounts.json account deploy --url http://127.0.0.1:5055/rpc --name my_account
+
+        To see account creation details, visit:
+        account: [..]
+        "},
+    );
+
+    let contents = fs::read_to_string(temp_dir.path().join(accounts_file))
+        .expect("Unable to read created file");
+
+    let expected = json!(
+        {
+            "alpha-sepolia": {
+                "my_account": {
+                    "address": "0x[..]",
+                    "class_hash": "0x[..]",
+                    "deployed": false,
+                    "legacy": false,
+                    "private_key": "0x[..]",
+                    "public_key": "0x[..]",
+                    "salt": "0x1",
+                    "type": "argent"
                 }
             }
         }
@@ -379,6 +447,56 @@ pub async fn test_happy_case_keystore(account_type: &str) {
 
     assert_matches(
         get_keystore_account_pattern(account_type.parse().unwrap(), None),
+        contents,
+    );
+}
+
+// TODO: Remove this test once we drop Argent account type
+#[tokio::test]
+pub async fn test_happy_case_keystore_argent_with_deprecation_warning() {
+    let temp_dir = tempdir().expect("Unable to create a temporary directory");
+    let keystore_file = "my_key.json";
+    let account_file = "my_account.json";
+    set_create_keystore_password_env();
+
+    let args = vec![
+        "--keystore",
+        keystore_file,
+        "--account",
+        account_file,
+        "account",
+        "create",
+        "--url",
+        URL,
+        "--type",
+        "argent",
+    ];
+
+    let snapbox = runner(&args).current_dir(temp_dir.path());
+
+    snapbox.assert().stdout_matches(formatdoc! {r"
+        [WARNING] Argent has rebranded as Ready. The `--argent` option for `account create` is now deprecated, please use `--ready` instead.
+        
+        Success: Account created
+
+        Address: 0x0[..]
+        
+        Account successfully created but it needs to be deployed. The estimated deployment fee is [..] STRK. Prefund the account to cover deployment transaction fee
+
+        After prefunding the account, run:
+        sncast --account {} --keystore {} account deploy --url {}
+
+        To see account creation details, visit:
+        account: [..]
+    ", account_file, keystore_file, URL});
+
+    assert!(temp_dir.path().join(keystore_file).exists());
+
+    let contents = fs::read_to_string(temp_dir.path().join(account_file))
+        .expect("Unable to read created file");
+
+    assert_matches(
+        get_keystore_account_pattern("argent".parse().unwrap(), None),
         contents,
     );
 }
@@ -708,6 +826,7 @@ pub async fn test_happy_case_default_name_generation() {
 fn get_formatted_account_type(account_type: &str) -> &str {
     match account_type {
         "oz" => "open_zeppelin",
+        "ready" => "argent",
         _ => account_type,
     }
 }
@@ -732,12 +851,12 @@ fn get_keystore_account_pattern(account_type: AccountType, class_hash: Option<&s
                 }
             )
         }
-        AccountType::Ready => {
+        AccountType::Argent | AccountType::Ready => {
             json!(
                 {
                     "version": 1,
                     "variant": {
-                        "type": "ready",
+                        "type": format!("{account_type}"),
                         "version": 1,
                         "owner": "0x[..]",
                         "guardian": "0x0"